# packages/contracts/generated/analyzer-openapi.json
{
  "components": {
    "schemas": {
      "AddressObservation": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "anyOf": [
              {
                "maxLength": 320,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Address"
          },
          "displayName": {
            "anyOf": [
              {
                "maxLength": 320,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Displayname"
          },
          "domain": {
            "anyOf": [
              {
                "maxLength": 253,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Domain"
          },
          "source": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Source",
            "type": "string"
          },
          "value": {
            "maxLength": 320,
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value",
          "source"
        ],
        "title": "AddressObservation",
        "type": "object"
      },
      "AnalysisFailure": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/AnalysisFailureCode"
          },
          "message": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "requestId": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Requestid"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "AnalysisFailure",
        "type": "object"
      },
...
      "AnalysisIntakeAccepted": {
        "additionalProperties": false,
        "properties": {
          "analysisRunId": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Analysisrunid",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatusValue",
            "default": "accepted"
          }
        },
        "required": [
          "analysisRunId"
        ],
        "title": "AnalysisIntakeAccepted",
        "type": "object"
      },
      "AnalysisIntakeRequest": {
        "additionalProperties": false,
        "examples": [
          {
            "analysisRunId": "run_01",
            "artifact": {
              "byteSize": 24831,
              "digestAlgorithm": "sha256",
              "objectKey": "organizations/org_01/cases/case_01/artifacts/artifact_01.eml",
              "sha256": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
            },
            "caseId": "case_01",
            "organizationId": "org_01",
            "requestedAt": "2026-01-01T00:00:00Z"
          }
        ],
        "properties": {
          "analysisRunId": {
            "maxLength": 160,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Analysisrunid",
            "type": "string"
          },
          "artifact": {
            "$ref": "#/components/schemas/Artifact"
          },
          "caseId": {
            "maxLength": 160,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Caseid",
            "type": "string"
          },
          "organizationId": {
            "maxLength": 160,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Organizationid",
            "type": "string"
          },
          "requestedAt": {
            "format": "date-time",
            "title": "Requestedat",
            "type": "string"
          }
        },
        "required": [
          "caseId",
          "organizationId",
          "analysisRunId",
          "artifact",
          "requestedAt"
        ],
        "title": "AnalysisIntakeRequest",
        "type": "object"
      },
      "AnalysisPhase": {
        "enum": [
          "queued",
          "fetching_evidence",
          "parsing",
          "extracting",
          "enriching",
          "scoring",
          "completed",
          "failed"
        ],
        "title": "AnalysisPhase",
        "type": "string"
      },
      "AnalysisResult": {
        "additionalProperties": false,
        "properties": {
          "addresses": {
            "items": {
              "$ref": "#/components/schemas/AddressObservation"
            },
            "maxItems": 100,
            "title": "Addresses",
            "type": "array"
          },
          "analysisRunId": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Analysisrunid",
            "type": "string"
          },
          "analysisVersion": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Analysisversion",
            "type": "string"
          },
          "analyzedAt": {
            "format": "date-time",
            "title": "Analyzedat",
            "type": "string"
          },
          "artifactByteSize": {
            "exclusiveMinimum": 0,
            "maximum": 50000000,
            "title": "Artifactbytesize",
            "type": "integer"
          },
          "artifactDigestAlgorithm": {
            "$ref": "#/components/schemas/DigestAlgorithm",
            "default": "sha256"
          },
          "artifactSha256": {
            "maxLength": 64,
            "minLength": 64,
            "pattern": "^[0-9a-fA-F]{64}$",
            "title": "Artifactsha256",
            "type": "string"
          },
          "authConflicts": {
            "items": {
              "$ref": "#/components/schemas/AuthConflictObservation"
            },
            "maxItems": 100,
            "title": "Authconflicts",
            "type": "array"
          },
          "authentication": {
            "items": {
              "$ref": "#/components/schemas/AuthenticationObservation"
            },
            "maxItems": 100,
            "title": "Authentication",
            "type": "array"
          },
          "caseId": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Caseid",
            "type": "string"
          },
          "confidence": {
            "maximum": 1,
            "minimum": 0,
            "title": "Confidence",
            "type": "number"
          },
          "containerSuspected": {
            "default": false,
            "title": "Containersuspected",
            "type": "boolean"
          },
          "contentIndicators": {
            "items": {
              "$ref": "#/components/schemas/ContentIndicatorObservation"
            },
            "maxItems": 1000,
            "title": "Contentindicators",
            "type": "array"
          },
          "dateObservations": {
            "items": {
              "$ref": "#/components/schemas/DateObservation"
            },
            "maxItems": 1000,
            "title": "Dateobservations",
            "type": "array"
          },
          "enrichment": {
            "items": {
              "$ref": "#/components/schemas/EnrichmentObservation"
            },
            "maxItems": 1000,
            "title": "Enrichment",
            "type": "array"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "maxItems": 500,
            "title": "Findings",
            "type": "array"
          },
          "headers": {
            "items": {
              "$ref": "#/components/schemas/HeaderObservation"
            },
            "maxItems": 1000,
            "title": "Headers",
            "type": "array"
          },
          "identityObservations": {
            "items": {
              "$ref": "#/components/schemas/IdentityObservation"
            },
            "maxItems": 100,
            "title": "Identityobservations",
            "type": "array"
          },
          "indicators": {
            "items": {
              "$ref": "#/components/schemas/IndicatorObservation"
            },
            "maxItems": 1000,
            "title": "Indicators",
            "type": "array"
          },
          "linkMismatches": {
            "items": {
              "$ref": "#/components/schemas/LinkMismatchObservation"
            },
            "maxItems": 1000,
            "title": "Linkmismatches",
            "type": "array"
          },
          "messageIdObservations": {
            "items": {
              "$ref": "#/components/schemas/MessageIdObservation"
            },
            "maxItems": 1000,
            "title": "Messageidobservations",
            "type": "array"
          },
          "mimeParts": {
            "items": {
              "$ref": "#/components/schemas/MimePartObservation"
            },
            "maxItems": 200,
            "title": "Mimeparts",
            "type": "array"
          },
          "nestedMessages": {
            "items": {
              "$ref": "#/components/schemas/NestedMessageObservation"
            },
            "maxItems": 10,
            "title": "Nestedmessages",
            "type": "array"
          },
          "organizationId": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Organizationid",
            "type": "string"
          },
          "parserWarnings": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "title": "Parserwarnings",
            "type": "array"
          },
          "receivedHops": {
            "items": {
              "$ref": "#/components/schemas/ReceivedHop"
            },
            "maxItems": 200,
            "title": "Receivedhops",
            "type": "array"
          },
          "routingAnomalies": {
            "items": {
              "$ref": "#/components/schemas/RoutingAnomalyObservation"
            },
            "maxItems": 200,
            "title": "Routinganomalies",
            "type": "array"
          },
          "rulesetVersion": {
            "default": "v1.2.0",
            "maxLength": 80,
            "minLength": 1,
            "title": "Rulesetversion",
            "type": "string"
          },
          "schemaVersion": {
            "default": "1.2.0",
            "maxLength": 20,
            "minLength": 1,
            "title": "Schemaversion",
            "type": "string"
          },
          "score": {
            "$ref": "#/components/schemas/ScoreBreakdown"
          },
          "verdict": {
            "$ref": "#/components/schemas/VerdictValue"
          }
        },
        "required": [
          "analysisVersion",
          "analysisRunId",
          "organizationId",
          "caseId",
          "artifactSha256",
          "artifactByteSize",
          "score",
          "verdict",
          "confidence",
          "analyzedAt"
        ],
        "title": "AnalysisResult",
        "type": "object"
      },
      "AnalysisStatus": {
        "additionalProperties": false,
        "properties": {
          "analysisRunId": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Analysisrunid",
            "type": "string"
          },
          "failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisFailure"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "phase": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisPhase"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "progress": {
            "anyOf": [
              {
                "maximum": 100,
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Progress"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatusValue"
          }
        },
        "required": [
          "analysisRunId",
          "status"
        ],
        "title": "AnalysisStatus",
        "type": "object"
      },
      "AnalysisStatusValue": {
        "enum": [
          "accepted",
          "queued",
          "processing",
          "completed",
          "deferred",
          "failed"
        ],
        "title": "AnalysisStatusValue",
        "type": "string"
      },
      "Artifact": {
        "additionalProperties": false,
        "properties": {
          "byteSize": {
            "exclusiveMinimum": 0,
            "maximum": 50000000,
            "title": "Bytesize",
            "type": "integer"
          },
          "digestAlgorithm": {
            "$ref": "#/components/schemas/DigestAlgorithm",
            "default": "sha256"
          },
          "objectKey": {
            "maxLength": 500,
            "minLength": 1,
            "pattern": "^organizations/[A-Za-z0-9_-]+/cases/[A-Za-z0-9_-]+/artifacts/[A-Za-z0-9_-]+\\.eml$",
            "title": "Objectkey",
            "type": "string"
          },
          "sha256": {
            "maxLength": 64,
            "minLength": 64,
            "pattern": "^[0-9a-fA-F]{64}$",
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "objectKey",
          "sha256",
          "byteSize"
        ],
        "title": "Artifact",
        "type": "object"
      },
...
      "ContainerFormat": {
        "enum": [
          "mbox",
          "bare_concatenation",
          "multipart/digest",
          "single"
        ],
        "title": "ContainerFormat",
        "type": "string"
      },
...
      "DigestAlgorithm": {
        "enum": [
          "sha256",
          "sha384",
          "sha512"
        ],
        "title": "DigestAlgorithm",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SegmentationRequest": {
        "additionalProperties": false,
        "properties": {
          "byteSize": {
            "exclusiveMinimum": 0,
            "maximum": 536870912,
            "title": "Bytesize",
            "type": "integer"
          },
          "caseId": {
            "maxLength": 160,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Caseid",
            "type": "string"
          },
          "evidenceId": {
            "maxLength": 160,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Evidenceid",
            "type": "string"
          },
          "objectKey": {
            "maxLength": 500,
            "minLength": 1,
            "pattern": "^organizations/[A-Za-z0-9_-]+/cases/[A-Za-z0-9_-]+/artifacts/[A-Za-z0-9_-]+(?:\\.[a-zA-Z0-9]+)?$",
            "title": "Objectkey",
            "type": "string"
          },
          "organizationId": {
            "maxLength": 160,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Organizationid",
            "type": "string"
          },
          "sha256": {
            "maxLength": 64,
            "minLength": 64,
            "pattern": "^[0-9a-fA-F]{64}$",
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "caseId",
          "organizationId",
          "evidenceId",
          "objectKey",
          "sha256",
          "byteSize"
        ],
        "title": "SegmentationRequest",
        "type": "object"
      },
      "SegmentationResult": {
        "additionalProperties": false,
        "properties": {
          "containerFormat": {
            "$ref": "#/components/schemas/ContainerFormat"
          },
          "messageCount": {
            "minimum": 0,
            "title": "Messagecount",
            "type": "integer"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/ContainerSegment"
            },
            "title": "Segments",
            "type": "array"
          }
        },
        "required": [
          "containerFormat",
          "messageCount"
        ],
        "title": "SegmentationResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerdictValue": {
        "enum": [
          "unknown",
          "benign",
          "suspicious",
          "malicious"
        ],
        "title": "VerdictValue",
        "type": "string"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "MailSentinel Analyzer",
    "version": "prototype-1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/analyses": {
      "post": {
        "operationId": "intake_v1_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisIntakeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisIntakeAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Intake"
      }
    },
    "/v1/analyses/{analysis_run_id}": {
      "get": {
        "operationId": "analysis_status_v1_analyses__analysis_run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_run_id",
            "required": true,
            "schema": {
              "title": "Analysis Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Analysis Status"
      }
    },
    "/v1/analyses/{analysis_run_id}/result": {
      "get": {
        "operationId": "analysis_result_v1_analyses__analysis_run_id__result_get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_run_id",
            "required": true,
            "schema": {
              "title": "Analysis Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Analysis Result"
      }
    },
    "/v1/evidence/segment": {
      "post": {
        "operationId": "segment_evidence_v1_evidence_segment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SegmentationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Segment Evidence"
      }
    }
  }
}